finished junit tests, beginning to implement add user through angular

+++ src/main/java/com/ajskubak/projectname/UserRepository.java
@@ -2,4 +2,4 @@ package com.ajskubak.projectname;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserModel,Integer> {}
+public interface UserRepository extends JpaRepository<UserModel,Long> {}
